requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
ruamel.yaml==0.18.6